Adds GLSL 120 hotfix.

+++ src/main/resources/assets/shaders/chunk_frag.glsl
@@ -59,7 +59,7 @@ varying vec3 normal;
 varying vec3 waterNormal;
 
 varying float flickeringLightOffset;
-varying int blockHint;
+varying float blockHint;
 varying float isUpside;
 
 void main(){
